+++ templates/slides/app/pages/DeckEditor.tsx
@@ -1129,6 +1129,7 @@ export default function DeckEditor() {
                   slidePresence={slidePresence}
                   recentEdits={deckRecentEdits}
                   aspectRatio={deck.aspectRatio}
+                  designSystem={designSystem}
                   generatingSlide={
                     generatingSlideVisible
                       ? {
@@ -1167,6 +1168,7 @@ export default function DeckEditor() {
               <GeneratingSlidePreview
                 content={streamedGeneratingSlideContent}
                 aspectRatio={deck.aspectRatio}
+                designSystem={designSystem}
                 thumbnail={false}
               />
             </div>
@@ -1182,6 +1184,7 @@ export default function DeckEditor() {
                 <GeneratingSlidePreview
                   content={streamedGeneratingSlideContent}
                   aspectRatio={deck.aspectRatio}
+                  designSystem={designSystem}
                   thumbnail={false}
                 />
               </div>